get multi tasking working with multicore

+++ cpu.c
@@ -1,12 +1,14 @@
 #include "cpu.h"
 #include "kerneldef.h"
+#include "tprintf.h"
 
 struct cpu cpus[2];
 proc cpu1_processes[1];
 proc cpu2_processes[2];
 
 void init_cpu(int max_processes, uint64 memory_area) {
-  int hart_id = cpuid();
+  uint64 hart_id = cpuid();
+
   struct cpu *cpu = &cpus[hart_id];
 
   cpu->processes = hart_id == 0 ? cpu1_processes : cpu2_processes;
